L0: Blackbox tests refactor [part IV]
This change adds a blackbox test for immediate command
list working in synchronous mode. It is intended to check
if append memory copy and append memory fill operations
work as expected.

Moreover, this change fixes a bug in cache implementation.

The goal is to split l0_test_basic into several smaller
blackbox tests.

Signed-off-by: Wrobel, Patryk <patryk.wrobel@intel.com>