fix: for zeCommandListHostSynchronizeHandler and USM registration
Provide custom handler, which performs `updatHostptrCopies()` in order
to correctly synchronize after `appendMemoryCopy()` is performed on an
immediate command list. Otherwise the execution could continue w/o
waiting for the copying to be completed.

Also, unregister USM buffer pointer once it is used to create proper
a memory descriptor to be transferred to client. Otherwise the container
of pointers keeps accumulating entries which may lead to future
collision of a valid pointer with an old entry.

This fixes a level-zero-test:
```
test_memory --gtest_filter='*AllocateMultipleDeviceMemoryAndThenFreeWith
DeferFreePolicyWhileBuffersInUse*'
```

Related-To: NEO-8910
Signed-off-by: Maciej Bielski <maciej.bielski@intel.com>