Print warning when exceeding user-available shmem
Add a thread-safe counter to track the shmem usage and compare that
with approximated total shmem amount available for the user. The total
amount is queried only at the allocator construction to minimize the
performance costs. Therefore, any dynamic changes happening in the
system while the service is active stay unnoticed.

Related-To: NEO-10624
Signed-off-by: Maciej Bielski <maciej.bielski@intel.com>